requirements.txt
torch>=1.10.0  
fair-esm>=2.0.0  
numpy>=1.19.0  
pandas>=1.1.0  
argparse  
scikit-learn>=0.24.0  
python>=3.7  
openpyxl>=3.0.0  